multiview/requirements.txt
# assume torch is already installed

# gsplat --index-url https://docs.gsplat.studio/whl/pt21cu121

# pycolmap for data parsing
git+https://github.com/rmbrualla/pycolmap@cc7ea4b7301720ac29287dbe450952511b32125e
# (optional) nerfacc for torch version rasterization 
# git+https://github.com/nerfstudio-project/nerfacc

viser
imageio[ffmpeg]
numpy==1.26.4
scikit-learn
tqdm
torchmetrics[image]
opencv-python
tyro
Pillow
tensorboard
